refactor: Alter NonTerminal merge functions signature to accept only NonTerminal nodes (#27)
The idea of this refactor is to use the type system to only allow
(un)ordered to be called on NonTerminal nodes, since calling them on
Terminal ones makes absolutely no sense. We also remove the possibility
of (un)ordered merge erroring because of trying to merge Terminal with
NonTerminal, letting us get rid of a test.