fix: Only consider argument types for matching (#33)
In the initial version of this rule, the name of the arguments were also
taken into account when calculating matching score for the signature of
the method. This is not precise and may lead into incorrect results
because argument names are not part of a method's signature, only the
argument types. So for instance: `void cadastrar(Pessoa umaPessoa)` and
`void cadastrar(Pessoa pessoa)` should match regardless of the Pessoa
argument name