fix(matching): Early return on matching if nodes have different kinds (#35)
This is a bug that was discovered thanks to running the tool in jdime
repository. Basically, the matching algorithm was running even if the
nodes have different kinds, which was introducing incorrect behavior in
the merge step: leading into trying merging two nodes that have
different kinds, which is logically inconsistent.