fix: Add id attribute in node (#42)
Since Rust uses structural equality to check if two structs are equal,
there was a bug in which two different nodes were being considered
equal. This introduced a bug in the matching data structure, causing the
scenario represented in
[bin/tests/scenarios/imports_are_merged_correctly](bin/tests/scenarios/imports_are_merged_correctly)
to crash. By generating a random UUID during parsing and using it for
checking equality between nodes, we drastically reduce the chances of
this issue occurring.