fix(matching): Incorrect calculation of perfect node matching (#58)
Previously, we were only running an equality check between both nodes
contents. Now, we rely on the property that two nodes are equal if their
matching score equals the sum of their trees' sizes. This makes the
check more accurate and more aligned with the strategies we're using in
the tool.